backend/requirements.txt
Flask==2.2.5
Flask-SocketIO==5.3.3
python-socketio==5.7.2
python-engineio==4.3.2
eventlet==0.33.3
SQLAlchemy==1.4.52
Flask-Cors==3.0.10
PyJWT==2.8.0
yfinance==0.2.27
python-dotenv==1.0.0
